split funny related functions to own file, move avatar to server
