+++ lib/index.js
@@ -1,6 +1,5 @@
 import Cleverbot from 'cleverbot.io';
 import gizoogle from 'gizoogle';
-import Imgflipper from 'imgflipper';
 import nconf from 'nconf';
 import R from 'ramda';
 import urban from 'urban';
@@ -11,11 +10,10 @@ import { mergeCommands } from './helpers';
 import sentry from './config/sentry';
 
 // Data
-import { choices, drama, eightball, emoji, memes, quotes } from '../data';
+import { choices, eightball, quotes } from '../data';
 
 // Init
 const clever = new Cleverbot(nconf.get('CLEVERBOT_API_NAME'), nconf.get('CLEVERBOT_API_KEY'));
-const imgflipper = new Imgflipper(nconf.get('IMGFLIP_USERNAME'), nconf.get('IMGFLIP_PASSWORD'));
 clever.setNick('Gravebot');
 
 let misc_commands = {
@@ -26,21 +24,6 @@ let misc_commands = {
       let rand = Math.floor(Math.random() * eightball.length);
       bot.sendMessage(msg.channel, `${msg.author}:crystal_ball:**${eightball[rand]}**:crystal_ball:`);
     }
-  },
-  avatar: (bot, msg, suffix) => {
-    if (msg.mentions.length === 0) {
-      if (msg.author.avatarURL === null) {
-        bot.sendMessage(msg.channel, 'You are naked.');
-      } else {
-        bot.sendMessage(msg.channel, `Your avatar:\n${msg.author.avatarURL}`);
-      }
-      return;
-    }
-    let msg_array = R.map(user => {
-      if (user.avatarURL === null) return `${user.username} is naked.`;
-      return `${user.username}\'s avatar:\n${user.avatarURL}`;
-    }, msg.mentions);
-    bot.sendMessage(msg.channel, msg_array);
   },
   chat: (bot, msg, suffix) => {
     if (!nconf.get('CLEVERBOT_API_NAME') || !nconf.get('CLEVERBOT_API_KEY')) {
@@ -78,45 +61,6 @@ let misc_commands = {
     } else {
       bot.sendMessage(msg.channel, 'Usage: **`!decide`** `something` **`or`** `something...`');
     }
-  },
-  drama: (bot, msg, suffix) => {
-    let rand = Math.floor(Math.random() * drama.length);
-    if (suffix && suffix >= 0 && suffix <= (drama.length - 1)) {
-      bot.sendMessage(msg.channel, drama[suffix]);
-    } else {
-      bot.sendMessage(msg.channel, drama[rand]);
-    }
-  },
-  emoji: (bot, msg, suffix) => {
-    let rand = Math.floor(Math.random() * emoji.length);
-    if (suffix && suffix >= 0 && suffix <= (emoji.length - 1)) {
-      bot.sendMessage(msg.channel, emoji[suffix]);
-    } else {
-      bot.sendMessage(msg.channel, emoji[rand]);
-    }
-  },
-  meme: (bot, msg, suffix) => {
-    if (!nconf.get('IMGFLIP_USERNAME') || !nconf.get('IMGFLIP_PASSWORD')) {
-      bot.sendMessage(msg.channel, 'Please setup imgflip in config.js to use the **`!meme`** command.');
-      return;
-    }
-
-    let usage_message = 'Usage: **`!meme`** `meme name` `"top text"` `"bottom text"`\nWrite **`!memelist`** for a list of meme names.';
-    if (!suffix) {
-      bot.sendMessage(msg.channel, usage_message);
-      return;
-    }
-    let tags = suffix.split('"');
-    let memetype = tags[0].trim();
-
-    imgflipper.generateMeme(memes[memetype], tags[1] ? tags[1] : '', tags[3] ? tags[3] : '', (err, image) => {
-      if (err) {
-        sentry.captureError(err);
-        bot.sendMessage(msg.channel, usage_message);
-      } else {
-        bot.sendMessage(msg.channel, image);
-      }
-    });
   },
   quote: (bot, msg, suffix) => {
     let rand = Math.floor(Math.random() * quotes.length);
