+++ lib/server.js
@@ -1,5 +1,22 @@
 import R from 'ramda';
 
+
+function avatar(bot, msg, suffix) {
+  if (msg.mentions.length === 0) {
+    if (msg.author.avatarURL === null) {
+      bot.sendMessage(msg.channel, 'You are naked.');
+    } else {
+      bot.sendMessage(msg.channel, `Your avatar:\n${msg.author.avatarURL}`);
+    }
+    return;
+  }
+  let msg_array = R.map(user => {
+    if (user.avatarURL === null) return `${user.username} is naked.`;
+    return `${user.username}\'s avatar:\n${user.avatarURL}`;
+  }, msg.mentions);
+  bot.sendMessage(msg.channel, msg_array);
+}
+
 function joinServer(bot, msg, suffix) {
   if (!suffix) {
     bot.sendMessage(msg.channel, 'Usage: **`!join`** `invitation link`');
@@ -82,6 +99,7 @@ ${uptimes} Seconds`);
 }
 
 export default {
+  avatar,
   join: joinServer,
   'join-server': joinServer,
   serverinfo,
